move walkDirAsync to file_utils.ts
