+++ util/file_utils.ts
@@ -15,3 +15,19 @@ export const walkDirSync = (dir: string, callback: (filename: string) => void): 
       callback(path.posix.join(dir, f));
   });
 };
+
+// Allows for an asynchronous callback, but still serializes them all
+// and does not run them in parallel.
+export const walkDirAsync = async (dir: string,
+    callback: (filename: string) => Promise<void>): Promise<void> => {
+  if (fs.statSync(dir).isFile()) {
+    await callback(path.posix.join(dir));
+    return;
+  }
+  const subdirs = fs.readdirSync(dir);
+  for (const subdir of subdirs) {
+    const dirPath = path.posix.join(dir, subdir);
+    const isDirectory = fs.statSync(dirPath).isDirectory();
+    await (isDirectory ? walkDirAsync(dirPath, callback) : callback(dirPath));
+  }
+};
